feat: enable machine pages functionality
This change allows optional page content to be created for individual
machines. This page content is rendered in the machine view, between the
inspector and the machine source.

Following the machine `slug` pattern, pages are created as
`machine-pages/{slug}.machine.tsx` and are automatically loaded for each
machine if present. If a page doesn't exist for a machine, nothing is
displayed in that space.

This change also includes a `MachinePageProps` interface which takes a
generic argument of the type of a machine. This allows machine pages to
take the machine service as the `service` prop with the correct types.